Switch back to random conn ID assignment, but use a u128 to make collisions irrelevant

+++ src/qos_net/src/error.rs
@@ -24,13 +24,13 @@ pub enum QosNetError {
 	/// DNS Resolution error
 	DNSResolutionError(String),
 	/// Attempt to save a connection with a duplicate connection ID
-	DuplicateConnectionId(u32),
+	DuplicateConnectionId(u128),
 	/// Error that should not happen: saving a connection overrode another
 	/// We take great care to ensure code prior to saving the connection checks
 	/// that it's not present in the connection map (and if it is, we return `DuplicateConnectionId`)
-	ConnectionOverridden(u32),
+	ConnectionOverridden(u128),
 	/// Attempt to send a message to a remote connection, but ID isn't found
-	ConnectionIdNotFound(u32),
+	ConnectionIdNotFound(u128),
 	/// Happens when a socket `read` returns too much data for the provided
 	/// buffer and the data doesn't fit. The first `usize` is the size of the
 	/// received data, the second `usize` is the size of the buffer.
